backend/pyproject.toml
[project]
name = "kleykod-backend"
version = "0.1.0"
description = "KleyKod - сервис объединения этикеток WB и Честного Знака"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "KleyKod Team" }
]

dependencies = [
    # Веб-фреймворк
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # Валидация и настройки
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",

    # Работа с файлами
    "pillow>=11.0.0",
    "pypdfium2>=4.30.0",
    "pylibdmtx>=0.1.10",
    "python-multipart>=0.0.18",

    # База данных
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",

    # Redis
    "redis>=5.2.0",

    # Шифрование (152-ФЗ)
    "cryptography>=44.0.0",

    # HTTP клиент
    "httpx>=0.28.0",

    # Утилиты
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=app --cov-report=term-missing"
